Improve encoding of subjects as strings
When there is a rfc that describes how to do something, it makes sense
to do it that way, especially since the old way had several flaws (e.g.
didn't escape reserved characters).

On the fence about what to do with the subject of the sets, as it looks
a little bit nicer with spaces in it, and it will never match the rfc
anyways (as we remove some fields). Keeping the old logic for now.