Improve error messages related to search type
Buypass doesn't include email address in newer (SEID 2) certs, so those
can't be found with email search, and that is nice to know for the user.
Also improve the somewhat confusing message when searching for
thumbprint - and only include it if we didn't find a match for the
thumbprint (the old behaviour was confusing as it gave the impression
that one couldn't trust the results even when we found a cert, which is
not the case).