Include the time in the pretty printed not_after
The certificates are often valid to 23:59 on a specific date, so it's
kinda misleading to only include the date, as it gives the impression
that it will expire almost 24h before it will. But it would not be
correct to round up either, so this is the best option, if somewhat
verbose.