Revert to old signature check
`verify_directly_issued_by(...)` does not support sha1, and we still
verify the signature on old sha1 certs, so that's a no go. Not sure we
still need to validate the signature on 10 year old certs that are long
expired, but that is a decision for another day. Just revert for now.