Set httpx+httpcore to one log level above ours
httpx now logs all requests at the INFO level, and that is a bit much in
prod when the brreg batch starts to run. But it's nice to have when
debugging, so set that log level to "one above" ours. And httpcore logs
some quite love level details on the DEBUG level, that is not really
necessary unless one is debugging some http problem, so do the same to
that.