Properly install ruldap3 in pipeline (#263)
The ruldap3 install was getting skipped if the venv came from cache,
which meant that we didn't test the ruldap3 in the current branch, but
the one from a earlier successfull build. This is bad. Force reinstall
to make sure we always end up with the one built from the current
branch.