Speed up testserver boot
Load duplicated certificates with keys from cache, if it exists, so that
we don't have to generate keys on every boot. Also, we don't actually
need to have real keys for the subscriber certs, as we don't use them
for anything. So just fake a public key instead of generating a real
private one.