Try to fix flaky API tests in CI (#282)
Uvicorn closes the connection after our 500 errors. So the theory is
that the httpx client in the api tests sometimes tries to reuse that
connection before it is actually shut down, and then ends up with a
ReadError. Add a matching `Connection: close` header so that httpx
(hopefully) understands that the connection is not for reuse.