Hopefully finally fix the Dependabot mess
Turns out that Dependabot runs with Python 3.12 and with the
`allow-unsafe` flag (if it finds an "unsafe" dependency already present)
so those changes where unnecessary. But if the "this is a scary
dependency" warning from pip-tools isn't there in the original
requirements file, it removes it, and the dependencies that follow it,
from the updates requirements file.

So let's reset the incorrect changes and add the adopt the requirements
from `uv` to match what makes Depandabot do the right thing.