Host our own static files
Hosting our own static files (instead of leaning on Caddy for this)
allows us to easier deploy to fly.io or something. It also keeps all the
config gathered here.

The static files should never change during the lifetime of the
container, so we don't need to check the disk for every request - just
get all the files at startup and cache them. And since both Starlette
and Granian supports pathsend, the file I/O is handled on the Rust side,
so it should be pretty fast.